+++ R/plot.R
@@ -1085,7 +1085,8 @@ plot_features_network <- function(ko_stat, map_id = "map00780",
                                   near_pathway = FALSE,
                                   modulelist = NULL,
                                   kos_color = c("Depleted" = "seagreen", "Enriched" = "orange", "None" = "grey", "Significant" = "red2", "Pathway" = "#80b1d3"),
-                                  pathway_label = TRUE, kos_label = TRUE,
+                                  pathway_label = TRUE, kos_label = TRUE, pathway_description = FALSE, kos_description = FALSE,
+                                  str_width = 50,
                                   mark_module = FALSE, mark_color = NULL,
                                   return_net = FALSE,
                                   ...) {
@@ -1137,7 +1138,22 @@ plot_features_network <- function(ko_stat, map_id = "map00780",
   igraph::vertex.attributes(ko_net)[["color"]] <- pcutils::tidai(igraph::vertex.attributes(ko_net)[["v_class"]], kos_color)
   tmp_v <- MetaNet::get_v(ko_net)
   if (!pathway_label) tmp_v$label <- ifelse(tmp_v$v_group == "Pathway", NA, tmp_v$label)
+  if (pathway_description) {
+    tmp_v$label <- ifelse(tmp_v$v_group == "Pathway", setNames(modulelist$Description,modulelist$id)[tmp_v$name], tmp_v$label) %>% stringr::str_wrap(., width = str_width)
+  }
   if (!kos_label) tmp_v$label <- ifelse(tmp_v$v_group == "KOs", NA, tmp_v$label)
+  if (kos_label & kos_description) {
+    ko_desc <- load_KO_desc()
+    if (grepl("C\\d{5}", tmp_v[tmp_v$v_group=="KOs","name"][1])) {
+      Compound_htable <- load_Compound_htable()
+      ko_desc <- Compound_htable[, c("Compound_id", "Compound_name")]
+      colnames(ko_desc) <- c("KO_id", "KO_name")
+    }
+    newname <- ko_desc[match(tmp_v[tmp_v$v_group=="KOs","name"], ko_desc$KO_id), "KO_name", drop = TRUE]
+    if (all(is.na(newname))) warning("No description for KO found, are you sure rownames of kodf are KOs?")
+    tmp_v[tmp_v$v_group=="KOs","label"] <- ifelse(is.na(newname), tmp_v[tmp_v$v_group=="KOs","name"], newname) %>% stringr::str_wrap(., width = str_width)
+  }
+
   if (mark_module) {
     ko_net_m <- MetaNet::module_detect(ko_net, method = "cluster_walktrap")
     if (return_net) {